+++ atlas.com/clc/rest/attributes/character.go
@@ -106,7 +106,7 @@ type CharacterSeedAttributesAttributes struct {
 	Hair      uint32 `json:"hair"`
 	HairColor uint32 `json:"hairColor"`
 	Skin      uint32 `json:"skin"`
-	Gender    uint32 `json:"gender"`
+	Gender    byte   `json:"gender"`
 	Top       uint32 `json:"top"`
 	Bottom    uint32 `json:"bottom"`
 	Shoes     uint32 `json:"shoes"`
